layer1/requirements.txt
google-generativeai
python-dotenv
pdf2image
Pillow
